ml-service/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
xgboost>=2.1.0
scikit-learn>=1.5.0
pandas>=2.2.0
numpy>=2.0.0
requests>=2.32.0
python-dotenv>=1.0.0
